Properly encode spaces in variable names for Conjur v5
Variable IDs were not encoded at all so in case a user had
a variable with spaces in it, we were not able to retrieve it.

The encoding method we use now encodes spaces into "%20" as required
by the Conjur v5 server. It also encodes slashes into "%2F" as required.

This commit updates the current UTs to expect "%2F" for slashes instead of
actual slashes (both are accepted in the server) and adds a UT for a variable
that has spaces in its name.