Fixed issue of not being able to find HFT-created Windows host creds
This code previously searched for windows credentials that did not match
saved ones when using host-factory tokens on windows (namely `<host>:<port>`),
causing Windows hosts to keep recreating their host identity with the
HF token. This would result in inability to fetch secrets once the HFT
token would expire but with this fix we find our host credentials in the
WinCred storage and can avoid recreating the host identity.