Fix build upgrades and remove `rspec-puppet` fork dependency
There were upstream issues with broken dependencies that we upstreamed a
fix for and because of this we needed to pin to latest (unreleased)
version of `rspec-puppet` gem. In the process, it seems like we also
discovered that our fork was not required as the tests all pass so it
was likely that we just had a fork to have a pre-release features ready
which are now in mainline.

Upstream bug: https://github.com/rodjek/rspec-puppet/issues/796
Upstream fix: https://github.com/rodjek/rspec-puppet/pull/797