Adds warning to set Conjur identity file permissions to 600
This change adds a warning to the main conjur-puppet README.md file that
the Conjur identity files `conjur.conf` and `conjur.identity` must have
their file permissions set to `600` when these files are manually
configured (pre-provisioned) on a Linux node.

This change also includes a modification to the
`examples/puppetmaster/smoketest_e2e.sh` script such that the file
permissions for the Conjur identity files for this smoketest are properly
set to `600`.

Addresses Issue #143