Heavily document `Sensitive` and `Deferred` requirements
Old documentation did not go into necessary depths to ensure that the
user was provided with all the necessary information about how the
`Deferred` and `Sensitive` chaining work and why they are critical in
ensuring secure use of credentials. This PR ensures that we do that and
fixes some outdated examples.